Add app controller for incident views and models
